recomp round 26: video plays, two root causes (sqrt bits, entry-first), boot 65 s -> 3 s
Two defects of our own runtime, both found by replaying a floor:

* recomp_fsqrt_f64 was declared double(double) while the lifter passes and
  expects bit patterns (recomp_wr64(dst, helper(recomp_rd64(src)))). Every
  positive square root came back a denormal, the game's sqrtf wrapper
  (0x00435a50) answered 0 for every vector length, and the door-touch check
  fired for the first open door in slot order every frame: the start-room
  ping-pong, the CellSpace::insert x1 > x2 grind and the inf/NaN entity
  positions were all that zero. Now uint64_t(uint64_t); sqrtsd/sqrtss are
  oracle cases; epochs 1700000000/1/2 do zero transitions and zero asserts
  before any input (the old "one transition" expectation was the defect
  stopping early).

* 813 goto-shaped lifted functions started at their lowest block instead of
  their entry. Ghidra bodies can absorb a block below the entry (a shared
  epilogue, a jump target); lift.py emitted blocks in address order. The
  string-table loader's second half (sub_00a27038) ran the loader's own
  epilogue and returned 0, so every '#KEY' stayed raw (#BASEMENT_NAME) since
  the first boot. lift.py now emits goto L_<entry> first; lift_patches.py
  --entry-first gives the already-lifted tree the same goto at build time
  (778 CRT-region thunks + 35 game functions, 11 TUs, no re-lift). The HUD
  reads "Basement" and "The Sad Onion".

Also in this round:

* Boot 65 s -> 2.9 s to the second frame: the archive mount's checksum pass
  is skipped by a block patch (ISAAC_ARCHIVE_VERIFY=1 restores it).
* Video plays: ISAAC_CUTSCENE=<frame>:<id> calls Manager::ShowCutscene from
  the frame present; libtheora's `emms; ret` body hand-written; the twelve
  SSE wide intrinsics that were aborting stubs implemented in recomp_rt.c
  and checked against Unicorn (the oracle compiles recomp_rt.c natively now).
* ISAAC_EPOCH pins the run RNG seed (replayable floors); ISAAC_HEAP_TRACE,
  ISAAC_DSP_TRACE, ISAAC_ROOM_PROBE/ISAAC_ROOM_TEST observe-only tools.
* scripts/recomp/web/drive_interactive.mjs: a state-driven browser driver
  (held keys) that starts a run with real key events and walks.
* scripts/recomp/assets/archive.py + optimize.py: the ARCH000 toolchain
  (verify/extract/pack/repack/shadow; lossless PNG + music q3 = the mounted
  set 1,070 -> 750 MB, staged in .scratch, not shipped yet).
* repentance.a is never named by this exe: unlisted from the lazy set;
  windowed-read selftest (203 checks, 0 warnings).

Verified after both fixes: node fast profile, epoch 1700000000, scripted
timeline -> "Basement", 3,000 frames, 0 transitions, 0 asserts, main 0;
headless Chromium fast module -> 3,000 frames in 90 s wall, "Basement",
100 PCM uploads / 16 plays, WebAudio running; interactive drive -> the run
starts at frame 633, walks, screenshots differ, ~35 fps, exit 0. The
optimised archive set boots the same way (verify: every entry matches).

Tests: recomp suites 105/105 (+ recomp-entry-first 4, recomp-fsqrt 3,
recomp-video 2, recomp-assets), host selftest 203/0, repo-safety 5/5.
Docs: recomp-architecture 21.40-21.41, HANDOFF, frontier machineTrack.

+++ scripts/recomp/lift/recomp_rt.h
@@ -293,7 +293,14 @@ static inline uint64_t recomp_f642bits(double d) { uint64_t b; memcpy(&b, &d, 8)
 RECOMP_DEF_FLT(f32, float, uint32_t, recomp_bits2f32, recomp_f322bits)
 RECOMP_DEF_FLT(f64, double, uint64_t, recomp_bits2f64, recomp_f642bits)
 
-double recomp_fsqrt_f64(double);           /* provided by recomp_rt.c */
+/* Both take and return the operand's BIT PATTERN, like every other f32/f64
+ * helper: the lifter emits recomp_wr64(dst, recomp_fsqrt_f64(recomp_rd64(src))).
+ * Round 26: the f64 one was declared double(double), so the bits arrived as a
+ * value and every positive root came back a denormal -- the game's sqrtf
+ * wrapper (0x00435a50) answered 0 for every vector length, every door within
+ * 25 px "touched", and positions went inf/NaN (the start-room ping-pong and
+ * the CellSpace grind). */
+uint64_t recomp_fsqrt_f64(uint64_t);       /* provided by recomp_rt.c */
 uint32_t recomp_fsqrt_f32(uint32_t);
 uint32_t recomp_fceil_f32(uint32_t);
 uint64_t recomp_fceil_f64(uint64_t);
